feat: config command
